implement timer

+++ app/src/main/java/com/furiousspider/oasisquiz/utils/Router.kt
@@ -20,7 +20,7 @@ object Router {
         activity.startActivity(intent)
     }
 
-    fun startSummaryActivity(activity: AppCompatActivity, score: Int? = null, time: Int? = null) {
+    fun startSummaryActivity(activity: AppCompatActivity, score: Int? = null, time: Double? = null) {
         val intent = Intent(activity, SummaryActivity::class.java)
         intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
         intent.putExtra(SummaryActivity.EXTRA_PREVIOUS_GAME, activity::class.java.name)
